Creating Timer component
