+++ src/components/quizz/Timer.css
@@ -1,0 +1,34 @@
+.chronometer {
+    position: absolute;
+    top: 29rem;
+    font-size: 3rem;
+    background-color: var(--blue);
+    font-weight: bold;
+    color: var(--white);
+    border-radius: 50%;
+    width: 8rem;
+    height: 8rem;
+    text-align: center;
+    margin: 0;
+    padding: 0;
+}
+
+.chronometer anim {
+    animation-name: time-counter;
+    animation-duration: 1s;
+}
+
+.chronometer p {
+    position: relative;
+    top: 2rem;
+    margin: 0;
+}
+
+@keyframes time-counter {
+    from {
+    }
+
+    to {
+        background-color: var(--orange);
+    }
+}
